BUGFIX: Don't skip a index job if a node can't be processed
The nodes that are used to create the index jobs are fetched from the database but hidden fields are not respected. The `NodeFactory` will right now return `null` for the configured context if the node is hidden. If that occurs the index job will return true and all other nodes in that job are skipped.

This fix will just skip the node that is hidden and log an information about that.